Refine auto build script: 1. Add time usage. 2. Skip precompile file. 3. Use os.remove to delete file. 4. Check if a file exists before deleting it.

+++ test/test_compile_each.py
@@ -4,9 +4,21 @@ import os
 import sys
 import shutil
 import subprocess
+import time
 
 CURRENT_DIRECTORY = os.getcwd()
 PARENT_DIRECTORY = os.pardir
+
+
+def TimePartsFromNs(duration):
+  min_part = duration // 1000000000 // 60
+  sec_part = duration // 1000000000 % 60
+  millisec_part = duration // 1000000 % 1000
+  return (min_part, sec_part, millisec_part)
+
+
+def FormatTimeNs(duration):
+  return '%d:%02d.%03d' % TimePartsFromNs(duration)
 
 
 def main():
@@ -14,16 +26,24 @@ def main():
   for file in os.listdir(PARENT_DIRECTORY):
     if not file.startswith('pe'):
       continue
+    filename, file_ext_name = os.path.splitext(file)
+    if file_ext_name != '':
+      continue
     content = '#include <%s>\n int main(){}' % file
     with open('main.cc', 'wb') as tempf:
       tempf.write(content.encode('utf8'))
     print('Compile %s' % file)
+    start_time = time.perf_counter_ns()
     ret = os.system('pe++.py main.cc')
-    os.remove('main.cc')
+    time_usage = FormatTimeNs(time.perf_counter_ns() - start_time)
+    if os.path.exists('main.cc'):
+      os.remove('main.cc')
+    print('Done, return code = %d, time usage = %s' % (ret, time_usage))
     print()
     if ret != 0:
       break
-  os.system('del a.exe')
+  if os.path.exists('a.exe'):
+    os.remove('a.exe')
   return ret
